Protect the use of document to resolve SSR issues.
To enable inclusion in applications that *may also use* SSR, the
invocation of `document.createRange()` needs to be moved out of the
global scope. Moving it into the `parseHTML()` method resolves this
issue since it will only be executed if the component is actually used
(on the browser side).

Full inclusion of this component by flowbite 2.4.0 introduced
SSR-related bugs across multiple frameworks (Nuxt, NextJS, Sveltekit)
due to this unprotected use of `document`.

Prior to flowbite 2.4.0, flowbite-datepicker was included via a "plugin"
approach that mitigated this unprotected use of `document`,

This resolves issue #41